added the contoller file for the newsFeed and removed that section out of the server and put it into that controller file. Added a userController file for after the user routes are tested and working okay. They will be moved there after.
