+++ client/src/component/Navbar.js
@@ -1,5 +1,10 @@
 import React from "react";
-import {Link} from "react-router-dom"
+import {Link} from "react-router-dom";
+import styled from 'styled-components';
+
+// const navBarBottom = styled.div`
+
+// `
 
 
 class  Navbar extends React.Component{
